backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.44
pymysql==1.1.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.10.1
python-multipart==0.0.20
reportlab==4.4.3
alembic==1.16.5
pytest==8.4.2
